+++ vscode/src/extension.ts
@@ -3,7 +3,7 @@ import { PanelManager } from './panel/PanelManager';
 import { CommandExecutor } from './commands/CommandExecutor';
 import { DiagnosticManager } from './diagnostics/DiagnosticManager';
 import { getFileInfo, parseLintResult, errorPositionToRange } from './utils/fileUtils';
-import { WebviewMessage, PanelState, DiagnosticType } from './utils/types';
+import { WebviewMessage, PanelState, DiagnosticType } from '../shared/types';
 
 let panelManager: PanelManager;
 let commandExecutor: CommandExecutor;
