+++ vscode/src/panel/PanelManager.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import * as fs from 'fs';
-import { WebviewMessage, PanelState } from '../utils/types';
+import { WebviewMessage, PanelState } from '../../shared/types';
 
 /**
  * Manages the webview panel lifecycle and content
